Agregada validacion sobre cantidad de semanas que un usuario puede comprar. Algún pequeño refactoring está en proceso...

+++ src/main/java/ar/edu/unlp/ttps/ComedorUNLP_TTPS_2015/controller/AdminController.java
@@ -29,7 +29,7 @@ public class AdminController {
 			@RequestParam("apellido") String apellido,
 			@RequestParam("dni") String dni,
 			@RequestParam("contrasena") String contrasena,
-			@RequestParam("telefono") Integer telefono,
+			@RequestParam("telefono") Long telefono,
 			@RequestParam("domicilio") String domicilio,
 			@RequestParam("email") String email) {
 
@@ -56,7 +56,7 @@ public class AdminController {
 	public ModelAndView editarAdmin(@RequestParam("id") Long id,
 			@RequestParam("nombre") String nombre,
 			@RequestParam("apellido") String apellido,
-			@RequestParam("telefono") Integer telefono,
+			@RequestParam("telefono") Long telefono,
 			@RequestParam("email") String email,
 			@RequestParam("telefono") String domicilio) {
